added time

+++ SUPER-MARIO-BROS-OOP/Project1/Game/GameConfig.cpp
@@ -71,7 +71,7 @@ void GameConfig::addCoin() {
 }
 
 // Set remaining time (seconds)
-void GameConfig::setTimeLeft(int seconds) {
+void GameConfig::setTimeLeft(float seconds) {
     if (seconds >= 0) {
         timeLeft = seconds;
     }
